#66: Moved and renamed mock server
